+++ cmd/block/info/process.go
@@ -17,6 +17,7 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
+	"net/http"
 	"os"
 	"strconv"
 	"strings"
@@ -55,11 +56,11 @@ func process(ctx context.Context, data *dataIn) (*dataOut, error) {
 		eth2Client: data.eth2Client,
 	}
 
-	specResponse, err := results.eth2Client.(eth2client.SpecProvider).Spec(ctx)
+	specResponse, err := results.eth2Client.(eth2client.SpecProvider).Spec(ctx, &api.SpecOpts{})
 	if err != nil {
 		return nil, errors.Wrap(err, "failed to connect to obtain configuration information")
 	}
-	genesisResponse, err := results.eth2Client.(eth2client.GenesisProvider).Genesis(ctx)
+	genesisResponse, err := results.eth2Client.(eth2client.GenesisProvider).Genesis(ctx, &api.GenesisOpts{})
 	if err != nil {
 		return nil, errors.Wrap(err, "failed to connect to obtain genesis information")
 	}
@@ -80,12 +81,14 @@ func process(ctx context.Context, data *dataIn) (*dataOut, error) {
 	})
 	if err != nil {
 		var apiErr *api.Error
-		if errors.As(err, &apiErr) && apiErr.StatusCode == 404 {
+		if errors.As(err, &apiErr) && apiErr.StatusCode == http.StatusNotFound {
 			if data.quiet {
 				os.Exit(1)
 			}
+
 			return nil, errors.New("empty beacon block")
 		}
+
 		return nil, errors.Wrap(err, "failed to obtain beacon block")
 	}
 	block := blockResponse.Data
@@ -345,7 +348,7 @@ func timeToBlockID(ctx context.Context, eth2Client eth2client.Service, input str
 	// Assume timestamp.
 	chainTime, err := standardchaintime.New(ctx,
 		standardchaintime.WithSpecProvider(eth2Client.(eth2client.SpecProvider)),
-		standardchaintime.WithGenesisTimeProvider(eth2Client.(eth2client.GenesisTimeProvider)),
+		standardchaintime.WithGenesisProvider(eth2Client.(eth2client.GenesisProvider)),
 	)
 	if err != nil {
 		return "", errors.Wrap(err, "failed to set up chaintime service")