+++ cmd/synccommitteemembers.go
@@ -31,7 +31,7 @@ var synccommitteeMembersCmd = &cobra.Command{
 In quiet mode this will return 0 if the synccommittee members are found, otherwise 1.
 
 epoch can be a specific epoch.  period can be 'current' for the current sync period or 'next' for the next sync period`,
-	RunE: func(cmd *cobra.Command, args []string) error {
+	RunE: func(cmd *cobra.Command, _ []string) error {
 		res, err := synccommitteemembers.Run(cmd)
 		if err != nil {
 			return err
@@ -49,7 +49,7 @@ epoch can be a specific epoch.  period can be 'current' for the current sync per
 func init() {
 	synccommitteeCmd.AddCommand(synccommitteeMembersCmd)
 	synccommitteeFlags(synccommitteeMembersCmd)
-	synccommitteeMembersCmd.Flags().Int64("epoch", -1, "the epoch for which to fetch sync committees")
+	synccommitteeMembersCmd.Flags().String("epoch", "", "the epoch for which to fetch sync committees")
 	synccommitteeMembersCmd.Flags().String("period", "", "the sync committee period for which to fetch sync committees ('current', 'next')")
 }